Add files via upload
Here is the Lenght Calculator, 
I tried to solve our basic problem in our archive: to calculate the lenght of the archives. We use most of the time 4 different boxes and each box has a different lenght. To calculate the total lenght is really important to estimate how much space is going to take. 
Instead of every time try to calculate with out basic smartphone calculators, I decided to write a program which is going to calculate automaticly for us. And it works!!